CN1-375 - Allow modules to be re-used regardless of where the tf files calling them are.
This would allow the Terragrunt modules to be re-usable regardless of the folder structure of the app using them.

The relative pathing is relative not to the path of the module, but to the folder structure of the GO/PackManager/CPI project.

For example, this works: "app_source/go/prepare.tf"
This doesn't: "app_source/go/another_folder/prepare.tf"

If we switch to absolute path, the modules are re-useable regardless of how the project structures their code.