Remove the skip variable from ecs modules
The usual workarounds for terraform's "maybe it exists, maybe it
doesn't" problems when switching between `count = 0` and `count = 1`
don't work when the output in question is a map type.

We receive a map type back from the new cloudformation that's used to
build the zero-downtime ecs cluster. So we were running into a lot of
issues where terraform couldn't tell whether it was supposed to be a
map, array, or string as the output (depending on what the `count` was
set to). Terraform currently doesn't support building an interpolation
that can work in all scenarios.

By elimating the possibility for two scenarios, we're able to eliminate
the error.

We've audited all use cases of the modules in question and found that
`skip` is not actually used anywhere, so it's safe to remove and simply
assume that the resources should always be provisioned.