CN1-700 - Fix conflict between autoscaling cluster and rollingupdate cluster.
During a rolling update, you must suspend Auto Scaling processes to avoid
making unexpected changes to the group. If an unexpected scaling action
changes the state of the group during a rolling update, this can cause
CloudFormation to have an inconsistent view of the group and can cause
the rolling update to fail. Processes to suspend include: HealthCheck,
ReplaceUnhealthy, AZRebalance, AlarmNotification, and ScheduledActions.

- This is a problem for CPI where we autoscale between 1-9 EC2 servers,
but also want to update the cluster regularly.
- This fix prevents random cluster rollbacks (which don't cause downtime).