DATAHOUSE-32: Piggyback on readonly role for data platform
Due to our huge number of account schemas, creating new grants on each
table in Ops is prone to breaking logical replication (at least while
using Postgres 11.5).  If we were to create a brand new custom role for
the data platform to use with a new set of grants, some of the queries
that debezium uses to initialize logical replication take extremely long
and cause debezium to drop its connection to the replication slot.